requirements.txt
opencv-python
flask
flask-wtf
pymysql
flask_sqlalchemy
flask-security-too
flask-mail
numpy
pytesseract
easyocr
imutils
gtts
pyspellchecker
indexer
pyvi
googletrans
ultralytics